move codegen for panic entry point to a separate mod
