+++ crates/rustc_codegen_spirv/src/builder/mod.rs
@@ -3,6 +3,7 @@ mod byte_addressable_buffer;
 mod ext_inst;
 mod intrinsics;
 pub mod libm_intrinsics;
+mod panic_entry_point;
 mod spirv_asm;
 
 pub use ext_inst::ExtInst;
